Refactors individual test suites into their own files
